requirement.txt
fastapi
uvicorn
boto3
python-dotenv
langchain-community
langchain-openai
langchain-huggingface
sentence-transformers
faiss-cpu
docx2txt
pypdf
python-docx
ollama
